Fix BNE and update memfile_3.dat
BNE did not have an opcode in the main decoder, so added the appropriate
opcode and flag translation

In our memfile_3.dat, the addi instruction to change $5 so it is not
equal to $4 was adding 0, updated it to add 1 as desired.

Added a simulation waveform config file to permit easy inclusion of the
desired parameters.